refactor(deck-views): rename mana_icon_service to mana_icon_factory; clean up pile view
- Rename widgets/mana_icon_service package to widgets/mana_icon_factory and
  update all importers.
- Hide the mana-cost/qty header above each pile and reclaim that top space.
- Draw stacked pile cards full-size in top-to-bottom order so each card fully
  covers the one beneath it, removing the transparent rounded-corner gaps
  that the previous strip-clipped rendering left between cards.
