+++ tests/test_mana_icon_factory.py
@@ -1,6 +1,6 @@
 import pytest
 
-from widgets.mana_icon_service import ManaIconFactory, normalize_mana_query, tokenize_mana_symbols
+from widgets.mana_icon_factory import ManaIconFactory, normalize_mana_query, tokenize_mana_symbols
 
 wx = pytest.importorskip("wx")
