+++ widgets/mana_icon_factory/cache.py
