+++ widgets/mana_icon_factory/resources.py
