+++ widgets/panels/card_inspector_panel/protocol.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any, Protocol
 
 import wx
 
-from widgets.mana_icon_service import ManaIconFactory
+from widgets.mana_icon_factory import ManaIconFactory
 from widgets.panels.card_image_display import CardImageDisplay
 from widgets.panels.mana_rich_text_ctrl import ManaSymbolRichCtrl
