+++ widgets/panels/deck_builder_panel/frame/search_results_view.py
@@ -13,7 +13,7 @@ from utils.constants import (
     BUILDER_NAME_COL_MIN_WIDTH,
     DARK_ALT,
 )
-from widgets.mana_icon_service import ManaIconFactory
+from widgets.mana_icon_factory import ManaIconFactory
 
 
 class _SearchResultsView(wx.ListCtrl):
@@ -57,7 +57,7 @@ class _SearchResultsView(wx.ListCtrl):
         Only costs absent from the cache require bitmap rendering, making
         repeated searches (including empty-filter / browse-all) O(new_costs).
         """
-        from widgets.mana_icon_service import tokenize_mana_symbols
+        from widgets.mana_icon_factory import tokenize_mana_symbols
 
         assert self._mana_icons is not None
         assert self._mana_img_list is not None
