+++ widgets/panels/deck_builder_panel/protocol.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any, Protocol
 
 import wx
 
-from widgets.mana_icon_service import ManaIconFactory
+from widgets.mana_icon_factory import ManaIconFactory
 
 if TYPE_CHECKING:
     from services.radar_service import RadarData
